+++ src/modules/mod.rs
@@ -82,6 +82,7 @@ pub mod threatfox;
 pub mod urlhaus;
 pub mod urlscan;
 pub mod username_search;
+pub mod username_variants;
 pub mod virustotal;
 pub mod waf_detect;
 pub mod wayback;
@@ -151,6 +152,7 @@ pub fn registry() -> Vec<Arc<dyn Module>> {
         Arc::new(employer_pivot::EmployerPivot),
         Arc::new(social_probe::SocialProbe),
         Arc::new(username_search::UsernameSearch),
+        Arc::new(username_variants::UsernameVariants),
         Arc::new(github_user::GithubUser),
         Arc::new(phone_intl::PhoneIntl),
         Arc::new(wayback::Wayback),